+++ packages/charts/chart-utilities/src/PlotlySchemaConverter.ts
@@ -1,4 +1,4 @@
-import type { Datum, TypedArray, PlotData, PlotlySchema, Data } from './PlotlySchema';
+import type { Datum, TypedArray, PlotData, PlotlySchema, Data, Layout, SankeyData } from './PlotlySchema';
 import { decodeBase64Fields } from './DecodeBase64Data';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
@@ -12,22 +12,19 @@ export interface OutputChartType {
   validTracesInfo?: [number, string][];
 }
 
-const SUPPORTED_PLOT_TYPES = [
-  'pie',
-  'bar',
-  'scatter',
-  'heatmap',
-  'sankey',
-  'indicator',
-  'gauge',
-  'histogram',
-  'histogram2d',
-];
-
 const UNSUPPORTED_MSG_PREFIX = 'Unsupported chart - type :';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
+export const isNumber = (value: any): boolean => !isNaN(parseFloat(value)) && isFinite(value);
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
 export const isDate = (value: any): boolean => {
+  // Don't consider number as date. There is no way to differentiate milliseconds from date and number
+  // without additional context.
+  if (isNumber(value)) {
+    return false;
+  }
+
   const parsedDate = new Date(Date.parse(value));
   if (isNaN(parsedDate.getTime())) {
     return false;
@@ -40,7 +37,13 @@ export const isDate = (value: any): boolean => {
   return true;
 };
 
-export const isNumber = (value: any): boolean => !isNaN(parseFloat(value)) && isFinite(value);
+const isYear = (input: string | number | Date | null): boolean => {
+  if (isNumber(input)) {
+    const possibleYear = typeof input === 'string' ? parseFloat(input) : Number(input);
+    return Number.isInteger(possibleYear) && possibleYear >= 1900 && possibleYear <= 2100;
+  }
+  return false;
+};
 
 export const isArrayOfType = (
   plotCoordinates: Datum[] | Datum[][] | TypedArray | undefined,
@@ -65,21 +68,22 @@ export const isArrayOfType = (
 };
 
 export const isDateArray = (data: Datum[] | Datum[][] | TypedArray | undefined): boolean => {
-  return isArrayOfType(data, isDate);
+  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+  return isArrayOfType(data, (value: any): boolean => isDate(value) || value === null);
 };
 
 export const isNumberArray = (data: Datum[] | Datum[][] | TypedArray | undefined): boolean => {
-  return isArrayOfType(data, isNumber);
+  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+  return isArrayOfType(
+    data,
+    (value: any): boolean =>
+      (typeof value === 'string' && isNumber(value)) || typeof value === 'number' || value === null,
+  );
 };
 
-export const isLineData = (data: Partial<PlotData>): boolean => {
-  return (
-    !SUPPORTED_PLOT_TYPES.includes(`${data.type}`) &&
-    Array.isArray(data.x) &&
-    isArrayOfType(data.y, (value: any) => typeof value === 'number') &&
-    data.x.length > 0 &&
-    data.x.length === data.y!.length
-  );
+export const isYearArray = (data: Datum[] | Datum[][] | TypedArray | undefined): boolean => {
+  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+  return isArrayOfType(data, (value: any): boolean => isYear(value) || value === null);
 };
 
 export const validate2Dseries = (series: Partial<PlotData>): boolean => {
@@ -91,6 +95,11 @@ export const validate2Dseries = (series: Partial<PlotData>): boolean => {
   }
 
   return true;
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const isInvalidValue = (value: any) => {
+  return typeof value === 'undefined' || value === null || (typeof value === 'number' && !isFinite(value));
 };
 
 const MAX_DEPTH = 15;
@@ -128,8 +137,11 @@ export const getValidSchema = (input: any): PlotlySchema => {
     if (!validatedSchema) {
       throw new Error('Plotly input is null or undefined');
     }
-    if (!validatedSchema.data) {
-      throw new Error('Plotly input data is null or undefined');
+    if (typeof validatedSchema !== 'object') {
+      throw new Error(`Plotly input is not an object. Input type: ${typeof validatedSchema}`);
+    }
+    if (!isArrayOrTypedArray(validatedSchema.data)) {
+      throw new Error('Plotly input data is not a valid array or typed array');
     }
     if (validatedSchema.data.length === 0) {
       throw new Error('Plotly input data is empty');
@@ -160,12 +172,81 @@ const validateBarData = (data: Partial<PlotData>) => {
 };
 
 const validateScatterData = (data: Partial<PlotData>) => {
-  if (data.mode === 'markers' && !isNumberArray(data.x)) {
-    throw new Error(`${UNSUPPORTED_MSG_PREFIX} ${data.type}, mode: ${data.mode}, xAxisType: String or Date`);
+  if (
+    ['markers', 'text+markers', 'markers+text'].includes(data.mode ?? '') &&
+    !isNumberArray(data.x) &&
+    !isDateArray(data.x)
+  ) {
+    throw new Error(`${UNSUPPORTED_MSG_PREFIX} ${data.type}, mode: ${data.mode}, xAxisType: String`);
   } else {
     validateSeriesData(data, true);
   }
 };
+
+const invalidateLogAxisType = (layout: Partial<Layout> | undefined): boolean => {
+  const isLogAxisType =
+    layout?.xaxis?.type === 'log' ||
+    layout?.yaxis?.type === 'log' ||
+    layout?.yaxis2?.type === 'log' ||
+    layout?.xaxis2?.type === 'log';
+
+  return isLogAxisType;
+};
+
+/**
+ * Detects cycles in Sankey chart data.
+ * @param nodes Array of node labels.
+ * @param links Array of links with source and target as node indices.
+ * @returns true if a cycle is found.
+ */
+function findSankeyCycles(input: Partial<SankeyData>): boolean {
+  const graph: Record<number, number[]> = {};
+  input.node?.label?.forEach((_, idx) => (graph[idx] = []));
+  input.link?.value?.forEach((val, idx) => {
+    if (
+      !(isInvalidValue(val) || isInvalidValue(input.link?.source?.[idx]) || isInvalidValue(input.link?.target?.[idx]))
+    ) {
+      graph[input.link!.source![idx]].push(input.link!.target![idx]);
+    }
+  });
+
+  const visited = new Set<number>();
+  const stack = new Set<number>();
+
+  function dfs(node: number, path: number[]) {
+    if (isInvalidValue(node) || !graph[node]) {
+      // Invalid node or no edges, return
+      return false;
+    }
+    if (stack.has(node)) {
+      // Cycle detected, return
+      return true;
+    }
+    if (visited.has(node)) {
+      return;
+    }
+
+    visited.add(node);
+    stack.add(node);
+    for (const neighbor of graph[node]) {
+      const cycleDetected = dfs(neighbor, [...path, neighbor]);
+      if (cycleDetected) {
+        return true; // Cycle found in the path
+      }
+    }
+    stack.delete(node);
+    return false; // No cycle found in this path
+  }
+
+  for (let i = 0; i < Object.keys(graph).length; i++) {
+    const cycleFound = dfs(i, [i]);
+    if (cycleFound) {
+      return true; // Cycle found
+    }
+  }
+
+  return false; // No cycles found
+}
 
 const DATA_VALIDATORS_MAP: Record<string, ((data: Data) => void)[]> = {
   indicator: [
@@ -176,12 +257,18 @@ const DATA_VALIDATORS_MAP: Record<string, ((data: Data) => void)[]> = {
     },
   ],
   histogram: [data => validateSeriesData(data as Partial<PlotData>, false)],
-  contour: [
+  bar: [
     data => {
-      throw new Error(`${UNSUPPORTED_MSG_PREFIX} ${data.type}`);
+      validateBarData(data as Partial<PlotData>);
     },
   ],
-  bar: [data => validateBarData(data as Partial<PlotData>)],
+  sankey: [
+    data => {
+      if (findSankeyCycles(data as Partial<SankeyData>)) {
+        throw new Error(`${UNSUPPORTED_MSG_PREFIX} ${data.type}, Cycles in Sankey chart not supported`);
+      }
+    },
+  ],
   scatter: [data => validateScatterData(data as Partial<PlotData>)],
   scatterpolar: [
     data => {
@@ -200,10 +287,7 @@ const getValidTraces = (dataArr: Data[]) => {
   const errorMessages: string[] = [];
   const validTraces = dataArr
     .map((data, index): [number, string] => {
-      let type = data.type;
-      if (isLineData(data as Partial<PlotData>)) {
-        type = 'scatter';
-      }
+      const type = data.type;
 
       if (type && DATA_VALIDATORS_MAP[type]) {
         const validators = DATA_VALIDATORS_MAP[type];
@@ -211,7 +295,7 @@ const getValidTraces = (dataArr: Data[]) => {
           try {
             validator(data);
           } catch (error) {
-            errorMessages.push(`data[${index}]: ${error}`);
+            errorMessages.push(`data[${index}] - type: ${data.type}, ${error}`);
             return [-1, DEFAULT_CHART_TYPE];
           }
         }
@@ -244,7 +328,12 @@ export const mapFluentChart = (input: any): OutputChartType => {
       return { isValid: false, errorMessage: `Failed to decode plotly schema: ${error}` };
     }
 
+    if (invalidateLogAxisType(validSchema.layout)) {
+      return { isValid: false, errorMessage: 'Log axis type is not supported' };
+    }
+
     const validTraces = getValidTraces(validSchema.data);
+
     const firstData = validSchema.data[validTraces[0][0]];
 
     switch (firstData.type) {
@@ -262,13 +351,11 @@ export const mapFluentChart = (input: any): OutputChartType => {
         return { isValid: true, type: 'verticalbar', validTracesInfo: validTraces };
       case 'scatterpolar':
         return { isValid: true, type: 'scatterpolar', validTracesInfo: validTraces };
+      case 'table':
+        return { isValid: true, type: 'table', validTracesInfo: validTraces };
       default:
         const containsBars = validTraces.some(trace => validSchema.data[trace[0]].type === 'bar');
-        const containsLines = validTraces.some(
-          trace =>
-            validSchema.data[trace[0]].type === 'scatter' ||
-            isLineData(validSchema.data[trace[0]] as Partial<PlotData>),
-        );
+        const containsLines = validTraces.some(trace => validSchema.data[trace[0]].type === 'scatter');
         if (containsBars && containsLines) {
           return { isValid: true, type: 'verticalstackedbar', validTracesInfo: validTraces };
         }
@@ -291,13 +378,19 @@ export const mapFluentChart = (input: any): OutputChartType => {
           });
           const isXDate = isDateArray(firstScatterData.x);
           const isXNumber = isNumberArray(firstScatterData.x);
-          if (isXDate || isXNumber) {
+          const isXYear = isYearArray(firstScatterData.x);
+          if ((isXDate || isXNumber) && !isXYear) {
             return { isValid: true, type: isAreaChart ? 'area' : 'line', validTracesInfo: validTraces };
+          } else if (isAreaChart) {
+            return {
+              isValid: false,
+              errorMessage: 'Fallback to VerticalStackedBarChart is not allowed for Area Charts.',
+            };
           }
           return { isValid: true, type: 'fallback', validTracesInfo: validTraces };
         }
 
-        return { isValid: false, errorMessage: `${UNSUPPORTED_MSG_PREFIX} :${firstData.type}` };
+        return { isValid: false, errorMessage: `${UNSUPPORTED_MSG_PREFIX} ${firstData.type}` };
     }
   } catch (error) {
     return { isValid: false, errorMessage: `Invalid plotly schema: ${error}` };