+++ packages/charts/react-charting/src/components/CommonComponents/CartesianChart.base.tsx
@@ -3,6 +3,7 @@ import { IProcessedStyleSet } from '@fluentui/react/lib/Styling';
 import { classNamesFunction, getId, getRTL } from '@fluentui/react/lib/Utilities';
 import { Callout } from '@fluentui/react/lib/Callout';
 import { FocusZone, FocusZoneDirection } from '@fluentui/react-focus';
+import { select as d3Select } from 'd3-selection';
 import {
   ICartesianChartStyles,
   ICartesianChartStyleProps,
@@ -11,7 +12,7 @@ import {
   IHorizontalBarChartWithAxisDataPoint,
   IHeatMapChartDataPoint,
 } from '../../index';
-import { convertToLocaleString } from '../../utilities/locale-util';
+import { formatToLocaleString } from '@fluentui/chart-utilities';
 import {
   createNumericXAxis,
   createStringXAxis,
@@ -30,6 +31,8 @@ import {
   ChartTypes,
   wrapContent,
   getSecureProps,
+  truncateString,
+  tooltipOfAxislabels,
 } from '../../utilities/index';
 import { LegendShape, Shape } from '../Legends/index';
 import { SVGTooltipText } from '../../utilities/SVGTooltipText';
@@ -89,6 +92,7 @@ export class CartesianChartBase
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   private _xScale: any;
   private isIntegralDataset: boolean = true;
+  private _tooltipId: string;
 
   constructor(props: IModifiedCartesianChartProps) {
     super(props);
@@ -104,6 +108,7 @@ export class CartesianChartBase
     this.idForGraph = getId('chart_');
     this.titleMargin = 8;
     this.idForDefaultTabbableElement = getId('defaultTabbableElement_');
+    this._tooltipId = getId('tooltip_');
     /**
      * In RTL mode, Only graph will be rendered left/right. We need to provide left and right margins manually.
      * So that, in RTL, left margins becomes right margins and viceversa.
@@ -222,14 +227,22 @@ export class CartesianChartBase
     points: any[],
     className: string,
   ): number => {
+    const formatTickLabel = (str: string) => {
+      if (this.props.showYAxisLablesTooltip) {
+        return truncateString(str, this.props.noOfCharsToTruncate || 4);
+      }
+
+      return str;
+    };
+
     if (chartType === ChartTypes.HeatMapChart) {
       return calculateLongestLabelWidth(
-        points[0].data.map((point: IHeatMapChartDataPoint) => point.y),
+        points[0].data.map((point: IHeatMapChartDataPoint) => formatTickLabel(`${point.y}`)),
         `.${className} text`,
       );
     } else {
       return calculateLongestLabelWidth(
-        points.map((point: IHorizontalBarChartWithAxisDataPoint) => point.y),
+        points.map((point: IHorizontalBarChartWithAxisDataPoint) => formatTickLabel(`${point.y}`)),
         `.${className} text`,
       );
     }
@@ -262,6 +275,15 @@ export class CartesianChartBase
     }
     // Callback for margins to the chart
     this.props.getmargins && this.props.getmargins(margin);
+
+    this._classNames = getClassNames(this.props.styles!, {
+      theme: this.props.theme!,
+      width: this.state._width,
+      height: this.state._height,
+      className: this.props.className,
+      isRtl: this._isRtl,
+      enableReflow: this.props.enableReflow,
+    });
 
     let callout: JSX.Element | null = null;
 
@@ -303,7 +325,7 @@ export class CartesianChartBase
       };
 
       const YAxisParams = {
-        margins: this.margins,
+        margins: this.props.getYDomainMargins ? this.props.getYDomainMargins(this.state.containerHeight) : this.margins,
         containerWidth: this.state.containerWidth,
         containerHeight: this.state.containerHeight - this.state._removalValueForTextTuncate!,
         yAxisElement: this.yAxisElement,
@@ -311,7 +333,7 @@ export class CartesianChartBase
         yAxisTickCount: this.props.yAxisTickCount!,
         yMinValue: this.props.yMinValue || 0,
         yMaxValue: this.props.yMaxValue || 0,
-        tickPadding: 10,
+        tickPadding: this.props.showYAxisLablesTooltip ? 15 : 10,
         maxOfYVal: this.props.maxOfYVal,
         yMinMaxValues: this.props.getMinMaxOfYAxis(points, this.props.yAxisType),
         // please note these padding default values must be consistent in here
@@ -447,20 +469,42 @@ export class CartesianChartBase
         );
       }
 
-      /*
-     * To create y axis tick values by if specified
-    truncating the rest of the text and showing elipsis
-    or showing the whole string,
-     * */
-      chartTypesToCheck.includes(this.props.chartType) &&
+      if (chartTypesToCheck.includes(this.props.chartType)) {
+        // To create y axis tick values by if specified truncating the rest of the text
+        // and showing elipsis or showing the whole string,
         yScalePrimary &&
-        createYAxisLabels(
-          this.yAxisElement,
-          yScalePrimary,
-          this.props.noOfCharsToTruncate || 4,
-          this.props.showYAxisLablesTooltip || false,
-          this._isRtl,
-        );
+          createYAxisLabels(
+            this.yAxisElement,
+            yScalePrimary,
+            this.props.noOfCharsToTruncate || 4,
+            this.props.showYAxisLablesTooltip || false,
+            this._isRtl,
+          );
+
+        // Removing un wanted tooltip div from DOM, when prop not provided, for proper cleanup
+        // of unwanted DOM elements, to prevent flacky behaviour in tooltips , that might occur
+        // in creating tooltips when tooltips are enabled( as we try to recreate a tspan with this._tooltipId)
+        if (!this.props.showYAxisLablesTooltip) {
+          try {
+            document.getElementById(this._tooltipId) && document.getElementById(this._tooltipId)!.remove();
+            //eslint-disable-next-line no-empty
+          } catch (e) {}
+        }
+        // Used to display tooltip at y axis labels.
+        if (this.props.showYAxisLablesTooltip) {
+          const yAxisElement = d3Select(this.yAxisElement).call(yScalePrimary);
+          try {
+            document.getElementById(this._tooltipId) && document.getElementById(this._tooltipId)!.remove();
+            //eslint-disable-next-line no-empty
+          } catch (e) {}
+          const ytooltipProps = {
+            tooltipCls: this._classNames.tooltip!,
+            id: this._tooltipId,
+            axis: yAxisElement,
+          };
+          yAxisElement && tooltipOfAxislabels(ytooltipProps);
+        }
+      }
 
       this.props.getAxisData && this.props.getAxisData(axisData);
       // Callback function for chart, returns axis
@@ -477,15 +521,6 @@ export class CartesianChartBase
         callout = this._generateCallout(calloutProps, chartHoverProps);
       }
     }
-
-    this._classNames = getClassNames(this.props.styles!, {
-      theme: this.props.theme!,
-      width: this.state._width,
-      height: this.state._height,
-      className: this.props.className,
-      isRtl: this._isRtl,
-      enableReflow: this.props.enableReflow,
-    });
 
     const svgDimensions = {
       width: this.state.containerWidth,
@@ -643,7 +678,7 @@ export class CartesianChartBase
                     this.state._removalValueForTextTuncate!,
                   y: this._isRtl
                     ? svgDimensions.width - this.margins.right! / 2 + this.titleMargin
-                    : this.margins.left! / 2 + this.state.startFromX - this.titleMargin,
+                    : this.margins.left! / 2 - this.titleMargin,
                   textAnchor: 'middle',
                   transform: `translate(0,
                    ${svgDimensions.height - this.margins.bottom! - this.margins.top! - this.titleMargin})rotate(-90)`,
@@ -719,7 +754,7 @@ export class CartesianChartBase
             className={this._classNames.calloutContentX}
             {...getAccessibleDataObject(calloutProps!.xAxisCalloutAccessibilityData, 'text', false)}
           >
-            {convertToLocaleString(calloutProps!.hoverXValue, this.props.culture)}
+            {formatToLocaleString(calloutProps!.hoverXValue, this.props.culture, this.props.useUTC)}
           </div>
         </div>
         <div
@@ -795,8 +830,8 @@ export class CartesianChartBase
       toDrawShape,
     });
 
-    const { culture } = this.props;
-    const yValue = convertToLocaleString(xValue.y, culture);
+    const { culture, useUTC } = this.props;
+    const yValue = formatToLocaleString(xValue.y, culture, useUTC);
     if (!xValue.yAxisCalloutData || typeof xValue.yAxisCalloutData === 'string') {
       return (
         <div style={yValueHoverSubCountsExists ? marginStyle : {}}>
@@ -818,9 +853,10 @@ export class CartesianChartBase
             <div>
               <div className={_classNames.calloutlegendText}> {xValue.legend}</div>
               <div className={_classNames.calloutContentY}>
-                {convertToLocaleString(
+                {formatToLocaleString(
                   xValue.yAxisCalloutData ? xValue.yAxisCalloutData : xValue.y ?? xValue.data,
                   culture,
+                  useUTC,
                 )}
               </div>
             </div>
@@ -837,9 +873,11 @@ export class CartesianChartBase
           {Object.keys(subcounts).map((subcountName: string) => {
             return (
               <div key={subcountName} className={_classNames.calloutBlockContainer}>
-                <div className={_classNames.calloutlegendText}> {convertToLocaleString(subcountName, culture)}</div>
+                <div className={_classNames.calloutlegendText}>
+                  {formatToLocaleString(subcountName, culture, useUTC)}
+                </div>
                 <div className={_classNames.calloutContentY}>
-                  {convertToLocaleString(subcounts[subcountName], culture)}
+                  {formatToLocaleString(subcounts[subcountName], culture, useUTC)}
                 </div>
               </div>
             );