+++ packages/charts/react-charting/src/utilities/SVGTooltipText.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { ITooltipHost, ITooltipProps, Tooltip, TooltipDelay } from '@fluentui/react/lib/Tooltip';
 import { Async, KeyCodes, getId, portalContainsElement } from '../Utilities';
-import { getRTL, ITheme } from '@fluentui/react';
+import { ITheme } from '@fluentui/react';
 
 interface ISVGTooltipTextProps {
   /**
@@ -123,7 +123,7 @@ export class SVGTooltipText
 
   public render(): React.ReactNode {
     const { content, tooltipProps, textProps, shouldReceiveFocus = true, showBackground = false } = this.props;
-    const { isTooltipVisible, textX, textY, textWidth, textHeight } = this.state;
+    const { isTooltipVisible, textWidth, textHeight } = this.state;
     const tooltipRenderProps: ITooltipProps = {
       content,
       targetElement: this._getTargetElement(),
@@ -140,14 +140,14 @@ export class SVGTooltipText
     const showTooltip =
       (!!this.props.isTooltipVisibleProp && this.state.isOverflowing && !!content) || (isTooltipVisible && !!content);
     const backgroundColor = this.props.theme ? this.props.theme.semanticColors.bodyBackground : undefined;
-    const isRTL = getRTL();
-    const rectX = isRTL ? (textX ?? 0) + (textWidth ?? 0) - PADDING : (textX ?? 0) - PADDING;
+    const rectX = (typeof textProps?.x === 'number' ? textProps.x : 0) - (textWidth ?? 0) / 2 - PADDING;
+    const rectY = (typeof textProps?.y === 'number' ? textProps.y : 0) - (textHeight ?? 0) / 2 - PADDING;
     return (
       <>
         {showBackground && (
           <rect
             x={rectX}
-            y={(textY ?? 0) - PADDING}
+            y={rectY}
             width={(textWidth ?? 0) + 2 * PADDING}
             height={(textHeight ?? 0) + 2 * PADDING}
             fill={backgroundColor}